0330
서버 수업 내용 정리
패킷 구조 업데이트

빌드까지는 정상적으로 동작하는 것을 확인했지만
실제로 패킷이 정상적으로 넘어가는지는 아직 테스트 하지 않음

+++ Source/HonorProject/Message/ServerAndClient.h
@@ -1,5 +1,8 @@
 #pragma once																										
-#include "GameServerMessage.h"																					
+#include "GameServerMessage.h"																	
+#include "MessageTypeEnum.h"
+#include "ContentsEnum.h"
+#include "ContentsStructure.h"
 																													
 class ChatMessage : public GameServerMessage											
 {																											
@@ -9,7 +12,7 @@ public:
 																												
 public:																										
 	ChatMessage()																		
-		: GameServerMessage(MessageType::Chat)											
+		: GameServerMessage(static_cast<uint32_t>(MessageType::Chat))					
 		, m_ID()																		
 		, m_Message()																		
 	{
